refector: Moved NoViewer components into separate folder
