+++ src/viewer/NoViewer/NoViewer.jsx
@@ -8,15 +8,15 @@ import CallToAction from './CallToAction'
 
 import { openLocalFileCopy } from 'drive/mobile/modules/offline/duck'
 
-import styles from './styles'
-import IconFiles from './icons/icon-type-files.svg'
-import IconBin from './icons/icon-type-bin.svg'
-import IconCode from './icons/icon-type-code.svg'
-import IconSpreadsheet from './icons/icon-type-sheet.svg'
-import IconSlide from './icons/icon-type-slide.svg'
-import IconText from './icons/icon-type-text.svg'
-import IconZip from './icons/icon-type-zip.svg'
-import IconPdf from './icons/icon-type-pdf.svg'
+import styles from '../styles.styl'
+import IconFiles from '../icons/icon-type-files.svg'
+import IconBin from '../icons/icon-type-bin.svg'
+import IconCode from '../icons/icon-type-code.svg'
+import IconSpreadsheet from '../icons/icon-type-sheet.svg'
+import IconSlide from '../icons/icon-type-slide.svg'
+import IconText from '../icons/icon-type-text.svg'
+import IconZip from '../icons/icon-type-zip.svg'
+import IconPdf from '../icons/icon-type-pdf.svg'
 
 class AsyncActionButton extends React.Component {
   state = {
